self-healing-agent/requirements.txt
# requirements.txt
Flask
flask-CORS
flask-Limiter
pymongo
requests
python-dotenv
schedule
redis
prometheus-flask-exporter
structlog
healthcheck
celery
pytest